first test version of cli build
